pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "duprly-mcp"
version = "0.1.0"
description = "MCP server for DUPR (Dynamic Universal Pickleball Rating) data management"
readme = "README_MCP.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DUPRLY Contributors"}
]
keywords = ["dupr", "pickleball", "mcp", "rating"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=0.9.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "SQLAlchemy>=2.0.4",
    "loguru>=0.6.0",
]

[project.scripts]
duprly-mcp = "duprly_mcp:main"

[project.optional-dependencies]
dev = [
    "flake8>=5.0.4",
    "pytest>=7.0.0",
]
sse = [
    "starlette>=0.27.0",
    "sse-starlette>=1.6.0",
    "uvicorn>=0.22.0",
]
keychain = [
    "keyring>=24.0.0",
]

[tool.setuptools]
py-modules = ["duprly_mcp", "dupr_client", "dupr_db", "dupr_resources", "duprly", "duprly_secrets"]

